requirements.txt
# Core dependencies
httpx>=0.28.0
click>=8.3.0
rich>=14.2.0
python-dotenv>=1.1.1

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Report conversion dependencies
markdown>=3.5.0
weasyprint>=60.0

# Development dependencies
pytest-mock>=3.12.0
